refactor(#132): extract event date/recurrence formatting; add Event Categories chip
- Extracted formatEventDateTime/formatRecurrence out of EventDetail.jsx
  into a new shared utils/eventSchedule.js (also exports
  getEventScheduleLine), so DirectionsModal.jsx can reuse the same
  formatting for its event date/schedule line without a circular
  import (EventDetail.jsx imports DirectionsModal, so the reverse
  isn't possible). EventDetail.jsx's own rendered output is unchanged.
- Added a clickable Categories entry to Event's "About + Details"
  (single-column render-function pattern, unlike the other 3 types'
  col1/col2 layout) — Events have a real 20-entry category taxonomy
  with live is_main assignments, matching the Business/Park/Trail
  Categories feature.

Refs #132, #139
