+++ nearby-app/app/src/components/details/EventDetail.jsx
@@ -6,7 +6,7 @@ import {
 } from 'lucide-react';
 
 import {
-  AccSection, QuickInfoRow, QuickInfoPhotosBox, SocialLinksGroup, hasSocialLinks,
+  AccSection, ContentGroup, CategoryChipList, QuickInfoRow, QuickInfoPhotosBox, SocialLinksGroup, hasSocialLinks,
   hasVal, copyToClipboard, getCoordinates, getImages,
 } from './shared';
 import InfoRow from './InfoRow';
@@ -17,6 +17,7 @@ import { EventJsonLd } from '../seo/index';
 import DirectionsModal from '../common/DirectionsModal';
 
 import { getDisplayableLocation } from '../../utils/getDisplayableLocation';
+import { formatEventDateTime, formatRecurrence } from '../../utils/eventSchedule';
 
 
 /* ------------------------------------------------------------------ */
@@ -34,47 +35,6 @@ const hasContent = (v) => {
   if (typeof v === 'object' && !Array.isArray(v) && Object.keys(v).length === 0) return false;
   return true;
 };
-
-function formatTime12h(d) {
-  if (!(d instanceof Date) || Number.isNaN(d.getTime())) return null;
-  const h24 = d.getHours();
-  const m = d.getMinutes();
-  const ampm = h24 >= 12 ? 'pm' : 'am';
-  const h = h24 % 12 === 0 ? 12 : h24 % 12;
-  return m === 0 ? `${h}${ampm}` : `${h}:${m.toString().padStart(2, '0')}${ampm}`;
-}
-
-/**
- * formatEventDateTime — always 12-hour. e.g. "Sat Nov 9th • 8am-7pm"
- */
-function formatEventDateTime(start, end) {
-  if (!start) return null;
-  const s = new Date(start);
-  if (Number.isNaN(s.getTime())) return null;
-
-  const dateStr = s.toLocaleDateString('en-US', {
-    weekday: 'short',
-    month: 'short',
-    day: 'numeric',
-  });
-  const day = s.getDate();
-  const suffix =
-    day % 10 === 1 && day !== 11 ? 'st'
-    : day % 10 === 2 && day !== 12 ? 'nd'
-    : day % 10 === 3 && day !== 13 ? 'rd' : 'th';
-  const dateWithOrdinal = dateStr.replace(/(\d+)$/, `$1${suffix}`);
-
-  const startTime = formatTime12h(s);
-  let endTime = null;
-  if (end) {
-    const e = new Date(end);
-    if (!Number.isNaN(e.getTime())) endTime = formatTime12h(e);
-  }
-
-  return endTime
-    ? `${dateWithOrdinal} • ${startTime}-${endTime}`
-    : `${dateWithOrdinal} • ${startTime}`;
-}
 
 /** Event-time derived open/closed label. */
 function deriveEventStatus(event) {
@@ -99,23 +59,6 @@ function formatCost(event) {
   if (typeof event.cost === 'string' && event.cost.trim() !== '') return event.cost;
   return null;
 }
-
-function formatRecurrence(event) {
-  if (!event?.is_repeating) return null;
-  const rp = event.repeat_pattern;
-  if (!rp) return 'This event repeats.';
-  if (typeof rp === 'string') return rp;
-  if (typeof rp === 'object') {
-    const parts = [];
-    if (rp.frequency) parts.push(rp.frequency);
-    if (rp.day_of_week) parts.push(`on ${rp.day_of_week}`);
-    if (rp.day_of_month) parts.push(`on day ${rp.day_of_month}`);
-    if (rp.interval) parts.push(`every ${rp.interval}`);
-    return parts.length > 0 ? `Repeats ${parts.join(', ')}` : 'This event repeats.';
-  }
-  return String(rp);
-}
-
 
 /* ------------------------------------------------------------------ */
 /* Component                                                           */
@@ -206,7 +149,8 @@ function EventDetail({ poi }) {
       hasContent(event?.organizer_email) ||
       hasContent(event?.organizer_phone) ||
       hasContent(event?.organizer_website);
-    if (!hasDesc && !hasTeaser && !recur && !hasOrganizer && !costLabel && !isCanceled && !hasContent(event?.status_explanation)) {
+    const hasCategories = Array.isArray(poi?.categories) && poi.categories.length > 0;
+    if (!hasDesc && !hasTeaser && !recur && !hasOrganizer && !costLabel && !isCanceled && !hasContent(event?.status_explanation) && !hasCategories) {
       return null;
     }
     return (
@@ -259,6 +203,9 @@ function EventDetail({ poi }) {
               </InfoRow>
             )}
           </div>
+        )}
+        {hasCategories && (
+          <ContentGroup title="Categories"><CategoryChipList categories={poi.categories} /></ContentGroup>
         )}
       </>
     );
